+++ aem-intellij-core/src/main/kotlin/com/aemtools/diagnostics/error/handler/provider/impl/GitHubAccessTokenHolder.kt
@@ -1,0 +1,33 @@
+package com.aemtools.diagnostics.error.handler.provider.impl
+
+import com.aemtools.diagnostics.error.handler.ErrorReportingConfig
+import com.aemtools.diagnostics.error.handler.crypto.TokenDecryptService
+import com.aemtools.diagnostics.error.handler.provider.AccessTokenHolder
+import java.nio.charset.StandardCharsets
+
+/**
+ * @author DeusBit
+ */
+class GitHubAccessTokenHolder : AccessTokenHolder {
+
+  override fun getToken(): String {
+    return GitHubAccessTokenProvider.token
+  }
+
+  object GitHubAccessTokenProvider {
+    private val config = ErrorReportingConfig()
+
+    val token: String by lazy {
+      val token = readToken()
+      String(TokenDecryptService.decrypt(token))
+    }
+
+    private fun readToken(): ByteArray {
+      return GitHubAccessTokenProvider::class.java.classLoader.getResource(config.tokenFile)
+          .readText(StandardCharsets.UTF_8)
+          .trim()
+          .toByteArray()
+    }
+  }
+}
+